feature: Make the syscall interface generic for unix targets
Use the libc crate and make the syscall interface work on any arch.

Also move unsafety to blocks and not whole functions.
