+++ smol_vm/Cargo.toml
@@ -5,3 +5,6 @@ version.workspace = true
 
 [dependencies]
 smol-file.workspace = true
+
+[target.'cfg(unix)'.dependencies]
+libc.workspace = true
