+++ smol_vm/src/syscall/mod.rs
@@ -1,4 +1,7 @@
-#[cfg(all(target_os = "linux", target_arch = "x86_64"))]
-pub mod x86_64_linux;
-#[cfg(all(target_os = "linux", target_arch = "x86_64"))]
-pub use x86_64_linux::vm_syscall;
+#[cfg(unix)]
+mod unix;
+
+#[cfg(unix)]
+use unix as imp;
+
+pub use imp::vm_syscall;
